:sparkles: add win capture animation
