+++ ft_gomoku/__init__.py
@@ -16,7 +16,8 @@ from .game_rules.capture import capture
 from .game_rules.ten_captures_to_win_rule import ten_capture_to_win
 from .game_rules.double_three_rule import double_three_forbidden
 from .data_structure import SettingsStruct
-from .engine import Engine
+from .engine import Engine, get_image, set_titlescreen, stop_sound, play_sound
+from .game import draw_rocks
 from .game import main_menu
 from .game import game_screen
 from .game import tutorial_screen
