+++ ft_gomoku/game/__init__.py
@@ -1,2 +1,3 @@
 from .screen import main_menu, game_screen, tutorial_screen
+from .screen import draw_rocks
 from .screen import Particle, stars_effect
