+++ ft_gomoku/game/screen/__init__.py
@@ -2,5 +2,6 @@ from .title_screen import main_menu
 from .game import game_screen
 from .tutorial import tutorial_screen
 from .particle import Particle, stars_effect
+from .components import draw_rocks
 from .square import Square
