+++ ft_gomoku/game/screen/animations/__init__.py
@@ -1,2 +1,2 @@
 from .place_rock import anim_place_rock
-from .win import anim_win, get_winner_rocks
+from .win import anim_win
