+++ ft_gomoku/game/screen/game.py
@@ -5,9 +5,10 @@ import pygame
 import time
 
 from ft_gomoku.AI.AI import run_ai
-from ft_gomoku.engine import Engine, get_image, set_titlescreen, play_sound, stop_sound
+from ft_gomoku import Engine, set_titlescreen, stop_sound
 from ft_gomoku.game.screen.components import draw_board, place_rocks, redraw_board
 from ft_gomoku.data_structure.GameStruct import GameStruct
+from ft_gomoku.game.screen.animations import anim_win
 from ft_gomoku import RuleStatus, five_to_win, double_three_forbidden, capture, ten_capture_to_win
 
 from ft_gomoku.data_structure.DebuggerStruct import DebuggerStruct
@@ -168,7 +169,7 @@ def game_screen(engine: Engine, ai: bool = False):
 				if result == 'play':
 					ai_rocks = None
 				if result == 'win':
-					win_screen(engine, game_engine, rocks_coord)
+					anim_win(engine, game_engine, rocks_coord)
 					return
 				if result == 'quit':
 					return
@@ -189,7 +190,7 @@ def game_screen(engine: Engine, ai: bool = False):
 					rocks_ia = run_ai(game_engine.grid, [double_three_forbidden, capture, ten_capture_to_win, five_to_win])
 					coords_to_place = (rocks_ia, rocks_coord[rocks_ia])
 					if place_rocks(engine.screen, game_engine, coords_to_place, False, 35) == 'win':
-						win_screen(engine, game_engine, rocks_coord)
+						anim_win(engine, game_engine, rocks_coord)
 						return
 					pass
 			if game_engine.grid.get_last_move() != game_engine.get_last_move(): # A CHANGER
@@ -204,26 +205,4 @@ def game_screen(engine: Engine, ai: bool = False):
 			pygame.display.update()
 
 
-def win_screen(engine: Engine, game_engine: GameStruct, rocks_coord: dict):
-	print('winner', game_engine.winner)
-	for x in range(30):
-		darken_screen(engine)
-		for i in range(5):
-			draw_rocks(engine.screen, game_engine, rocks_coord[game_engine.winner[1][i]], 35, game_engine.winner[0])
-			pygame.display.update()
-	play_sound('winning_sound.mp3')
-	for i in range(5):
-		draw_rocks(engine.screen, game_engine, rocks_coord[game_engine.winner[1][i]], 35, game_engine.winner[0], True)
-		pygame.display.update()
-		pygame.time.wait(100)
-	pygame.time.wait(1000)
-	engine.change_screen('main_menu')
-	return
 
-
-def darken_screen(engine: Engine):
-	overlay = pygame.Surface((engine.screen.get_width(), engine.screen.get_height()))  # Create a new surface with screen dimensions
-	overlay.set_alpha(15)
-	overlay.fill((0, 0, 0))
-	engine.screen.blit(overlay, (0, 0))
-	pygame.display.flip()
